[KDESKTOP-1199] Defines a log-free normalizedSyncName function in libcommon/utility (to be used in gui)
