+++ src/libcommon/CMakeLists.txt
@@ -138,7 +138,11 @@ if(APPLE)
     target_link_libraries(${libcommon_NAME}
         ${FOUNDATION_LIBRARY}
         ${CORESERVICES_LIBRARY}
-        ${APPKIT_LIBRARY})
+        ${APPKIT_LIBRARY} utf8proc)
+endif()
+
+if(UNIX AND NOT APPLE)
+    target_link_libraries(${libcommon_NAME} utf8proc)
 endif()
 
 if(ZLIB_FOUND)
