+++ src/server/requests/serverrequests.cpp
@@ -1504,7 +1504,7 @@ ExitCode ServerRequests::addSync(int driveDbId, const QString &localFolderPath, 
 #ifdef __APPLE__
     // On macOS, the special characters in file names are NFD encoded. However, we use QFileDialog::getExistingDirectory to
     // retrieve the selected sync path which return a NFC encoded path.
-    (void) Utility::normalizedSyncPath(localFolderPath.toStdString(), localPath, Utility::UnicodeNormalization::NFD);
+    (void) Utility::normalizedSyncPath(localFolderPath.toStdString(), localPath, UnicodeNormalization::NFD);
 #endif
     sync.setLocalPath(localPath);
     sync.setTargetPath(QStr2Path(serverFolderPath));
